Add files via upload
The linear regression model demonstrated excellent performance, with a high R-squared value (0.967) indicating that the model explains a significant portion of the variance in writing scores. The low MSE (106.56) further confirms the accuracy of the predictions. The scatter plot of actual vs predicted writing scores showed that most points are close to the 45-degree line validating the model's predictive power.